Fix dangling uintptr bugs
A uintptr is opaque to the runtime and so any time an unsafe.Pointer is
converted into a uintptr we must ensure that the underlying pointer
cannot be invalidated.

The documentation for
[unsafe.Pointer](https://pkg.go.dev/unsafe#Pointer) lists the safe ways
to use uintptr. The relevant subsection for this patch is 2), a uintptr has no pointer semantics and it is unsafe to treat it as a live reference except in a limited set of cases. Subsection 4) describes the exception we rely on, that converting unsafe.Pointer to uintptr in an argument is live until the call completes. We'll also store unsafe.Pointer instead of uintptr in the ifreq struct. Both types have the same underlying representation and so this is equivalent to the kernel.

+++ ethtool_link_settings.go
@@ -74,7 +74,7 @@ func (e *Ethtool) GetLinkSettings(intf string) (*LinkSettings, error) {
 	// Provide the maximum expected nwords based on our constant
 	req.Settings.LinkModeMasksNwords = int8(MAX_LINK_MODE_MASK_NWORDS)
 
-	err := e.ioctl(intf, uintptr(unsafe.Pointer(&req)))
+	err := e.ioctl(intf, unsafe.Pointer(&req))
 	fallbackReason := ""
 
 	var errno syscall.Errno
@@ -137,7 +137,7 @@ func (e *Ethtool) SetLinkSettings(intf string, settings *LinkSettings) error {
 	checkReq.Settings.Cmd = ETHTOOL_GLINKSETTINGS
 	checkReq.Settings.LinkModeMasksNwords = int8(MAX_LINK_MODE_MASK_NWORDS)
 
-	errGLinkSettings := e.ioctl(intf, uintptr(unsafe.Pointer(&checkReq)))
+	errGLinkSettings := e.ioctl(intf, unsafe.Pointer(&checkReq))
 	canUseGLinkSettings := false
 	nwords := 0
 
@@ -179,7 +179,7 @@ func (e *Ethtool) SetLinkSettings(intf string, settings *LinkSettings) error {
 		copy(setReq.Masks[0*nwords:1*nwords], zeroMaskSupported)
 		copy(setReq.Masks[2*nwords:3*nwords], zeroMaskLp)
 
-		if err := e.ioctl(intf, uintptr(unsafe.Pointer(&setReq))); err != nil {
+		if err := e.ioctl(intf, unsafe.Pointer(&setReq)); err != nil {
 			return fmt.Errorf("ETHTOOL_SLINKSETTINGS ioctl failed: %w", err)
 		}
 		return nil
